Add files via upload
Edge Impulse ML model to classify tremor frequency based on ax, ay, az values from an accelerometer.

The code is for an Arduino sketch that uses the Edge Impulse ingestion SDK to classify data from an accelerometer. The sketch first initializes the serial port and the IMU, and then it starts sampling data from the IMU. The data is then converted to a signal and passed to the classifier. The classifier then outputs a prediction, which is printed to the serial port.

The code uses the following libraries:

TremorDuino_inferencing.h: This library provides the functions for initializing the IMU and for converting the data from the IMU to a signal.
Arduino_LSM9DS1.h: This library provides the functions for reading data from the IMU.
The code uses the following constants:

CONVERT_G_TO_MS2: This constant is used to convert the acceleration data from the IMU to meters per second squared.
MAX_ACCEPTED_RANGE: This constant is used to limit the range of the acceleration data.
The code uses the following functions:

ei_get_sign(): This function returns the sign of a number.
numpy::signal_from_buffer(): This function creates a signal from a buffer of data.
run_classifier(): This function runs the classifier on a signal.
The code works as follows:

The sketch initializes the serial port and the IMU.
The sketch starts sampling data from the IMU.
The data is converted to a signal.
The signal is passed to the classifier.
The classifier outputs a prediction.
The prediction is printed to the serial port.
The sketch can be used to classify data from any accelerometer. However, the accuracy of the classification will depend on the quality of the data and the accuracy of the model.

Here are some additional notes about the code:

The code uses the -DEI_CLASSIFIER_ALLOCATION_STATIC compiler flag to allocate the classifier memory statically. This can help to improve performance on devices with limited memory.
The code uses the debug_nn variable to control whether or not debug information is printed to the serial port.
The code uses the #if preprocessor directives to check for the presence of certain macros. This can help to prevent errors if the code is used with a model that does not support a particular feature.


The prediction value is stored in the result.classification array. This array contains a list of ei_classification_result_t structures, each of which represents a possible classification. The value field of each structure contains the probability of that classification. The highest probability classification is the prediction.

Here is an example of how to print the prediction value:

ei_printf("Prediction: %s (%.5f)\n", result.classification[0].label, result.classification[0].value);